Pass WITTY_MODULE_HASH to source
This allows source code to generate typenames based on the witty module
hash. This is needed when using nanobind to generate multiple
instantiations of the same class (nanobind keeps a global registry and
will not allow two types with the same typename to co-exist, even if
compiled in different modules).